requirements.txt
amqp==2.3.2
appdirs==1.4.3
Babel==2.4.0
billiard==3.5.0.4
cairocffi==0.8.0
CairoSVG==2.1.1
celery==4.2.1
cffi==1.11.2
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.10.6
django-admin-honeypot==1.1.0
django-allauth==0.31.0
django-auditlog==0.4.3
django-bootstrap-datepicker==1.2.2
django-bootstrap-form==3.2.1
django-bootstrap3==8.2.1
django-bootstrap3-datetimepicker-2==2.4.2
django-cleanup==1.0.1
django-datetime-widget==0.9.3
django-dynamic-formsets==0.0.8
django-jsonfield==1.0.1
django-mailer==1.2.5
django-multiupload==0.5.2
django-phonenumber-field==1.3.0
django-post-office==3.0.3
django-tinymce==2.6.0
django-unused-media==0.1.9
jsonfield==2.0.2
kombu==4.2.1
lockfile==0.12.2
mysqlclient==1.3.10
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pdfrw==0.4
phonenumberslite==8.4.2
Pillow==4.0.0
pycparser==2.18
pyparsing==2.2.0
Pyphen==0.9.4
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
tinycss2==0.6.1
uWSGI==2.0.15
vine==1.1.4
WeasyPrint==0.41
webencodings==0.5.1
amqp==2.3.2
appdirs==1.4.3
Babel==2.4.0
billiard==3.5.0.4
cairocffi==0.8.0
CairoSVG==2.1.1
celery==4.2.1
cffi==1.11.2
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-auditlog==0.4.3
django-bootstrap-datepicker==1.2.2
django-bootstrap-form==3.2.1
django-bootstrap3==8.2.1
django-bootstrap3-datetimepicker-2==2.4.2
django-cleanup==1.0.1
django-datetime-widget==0.9.3
django-dynamic-formsets==0.0.8
django-jsonfield==1.0.1
django-multiupload==0.5.2
django-phonenumber-field==1.3.0
django-post-office==3.0.3
django-tinymce==2.6.0
django-unused-media==0.1.9
#html5lib==1.0.1
jsonfield==2.0.2
kombu==4.2.1
mysqlclient==1.3.10
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pdfrw==0.4
phonenumberslite==8.4.2
Pillow==4.0.0
pycparser==2.18
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
tinycss2==0.6.1
uWSGI==2.0.15
vine==1.1.4
WeasyPrint==0.41
webencodings==0.5.1
xhtml2pdf==0.2.3
amqp==2.3.2
appdirs==1.4.3
Babel==2.4.0
billiard==3.5.0.4
cairocffi==0.8.0
CairoSVG==2.1.1
celery==4.2.1
cffi==1.11.2
cssselect2==0.2.1
defusedxml==0.5.0
Django==1.10.6
django-admin-honeypot==1.1.0
django-allauth==0.31.0
django-auditlog==0.4.3
django-bootstrap-datepicker==1.2.2
django-bootstrap-form==3.2.1
django-bootstrap3==8.2.1
django-bootstrap3-datetimepicker-2==2.4.2
django-cleanup==1.0.1
django-datetime-widget==0.9.3
django-dynamic-formsets==0.0.8
django-jsonfield==1.0.1
django-multiupload==0.5.2
django-phonenumber-field==1.3.0
django-post-office==3.0.3
django-tinymce==2.6.0
django-unused-media==0.1.9
html5lib==1.0.1
jsonfield==2.0.2
kombu==4.2.1
mysqlclient==1.3.10
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pdfrw==0.4
phonenumberslite==8.4.2
Pillow==4.0.0
pycparser==2.18
pyparsing==2.2.0
PyPDF2==1.26.0
Pyphen==0.9.4
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
reportlab==3.4.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
tinycss2==0.6.1
uWSGI==2.0.15
vine==1.1.4
WeasyPrint==0.41
webencodings==0.5.1
xhtml2pdf==0.2.3
